+++ src/services/unique/HyperUniqueField.php
@@ -1,0 +1,52 @@
+<?php
+/**
+ * Craft CMS Plugins
+ *
+ * Created with PhpStorm.
+ *
+ * @link      https://github.com/Anubarak/
+ * @email     anubarak1993@gmail.com
+ * @copyright Copyright (c) 2024 Robin Schambach|Secondred Newmedia GmbH
+ */
+
+namespace anubarak\seeder\services\unique;
+
+use anubarak\seeder\Seeder;
+use craft\base\Field;
+
+/**
+ * Class HyperUniqueField
+ *
+ * @package anubarak\seeder\services\unique
+ * @since   10.07.2024
+ * @author  by Robin Schambach
+ */
+class HyperUniqueField implements UniqueFieldInterface
+{
+    /**
+     * @inheritDoc
+     */
+    public function getDescription(Field $field): string
+    {
+        return 'Link value | empty';
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getValues(Field $field): array
+    {
+        return [
+            null,
+            fn() => Seeder::$plugin->getSeeder()->getFieldData($field)
+        ];
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getFieldClass(): string
+    {
+        return 'verbb\\hyper\\fields\\HyperField';
+    }
+}